requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.16.0
webdriver-manager==4.0.1
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.1
python-dotenv==1.0.0
schedule==1.2.1
lxml==5.1.0
fake-useragent==1.4.0
